Add 'Start Game' button to begin game. Close #5
- Add CSS and HTML to render a welcome message with 'Start Game' button
- Pull out game logic from index.js into game.js
