--- css/application.css
@@ -1,14 +1,0 @@
-body {
-  background-image: url('http://i.imgur.com/CAbj8La.jpg');
-  background-attachment: fixed;
-  background-size: cover;
-}
-
-.game-canvas {
-  background-color: rgba(0, 0, 0, 0.8);
-  border: 2px solid;
-  border-color: white;
-  display: block;
-  margin: auto;
-  margin-top: 5%;
-}
